+++ lib/money_extensions.rb
@@ -2,7 +2,7 @@ module MoneyExtensions
   require 'money'
   require 'money_extensions/array'
   require 'money_extensions/big_decimal'
-  require 'money_extensions/fixnum'
+  require 'money_extensions/integer'
   require 'money_extensions/extensions'
   require 'money_extensions/money_field'
   require 'money_extensions/numeric'