+++ lib/money_extensions/integer.rb
@@ -1,4 +1,4 @@
-class Fixnum
+class Integer
   #Returns self as a money (treated as cents)
   def total_money
     zero? ? ::Money.new(0) : ::Money.new(self)