+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+* Using Fixnum is deprecated
+
 ## 0.1.0
 
 * Drop rails 2 and old ruby support